+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jthuy <jthuy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 12:38:44 by jthuy             #+#    #+#             */
-/*   Updated: 2019/12/07 17:07:22 by jthuy            ###   ########.fr       */
+/*   Updated: 2019/12/07 17:19:56 by jthuy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,11 +63,11 @@ void	binto_oct(void **value, unsigned int *flags);
 // void	binto_hex(void **value, char index);
 void	binto_hex(void **value, unsigned int *flags, char index);
 // void	def_bitborder(unsigned int *bitborder, int value, char bit_count);
-void	def_bitborder(unsigned long long int *bitborder, void **value, unsigned int *flags);
+void	def_bitborder(long int *bitborder, void **value, unsigned int *flags);
 
-void	def_bitborder_int(unsigned long long int *bitborder, int value, char bit_count);
+void	def_bitborder_int(long int *bitborder, int value, char bit_count);
 
-void	def_bitborder_lint(unsigned long int *bitborder, long int value, char bit_count);
+void	def_bitborder_lint(long int *bitborder, long int value, char bit_count);
 
 /*
 ** put_numbers.c
